requirements.txt
# FastAPI core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Base de datos
psycopg2-binary==2.9.9
redis==5.0.1

# Autenticación
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilidades básicas
pydantic>=2.8.0
python-dateutil==2.8.2
aiofiles==23.2.1

# Machine Learning (CPU optimizado para Railway)
torch==2.1.0+cpu -f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.16.0+cpu -f https://download.pytorch.org/whl/torch_stable.html
ftfy==6.1.1
regex==2023.10.3
tqdm==4.66.1

# CLIP
git+https://github.com/openai/CLIP.git

# Procesamiento de imágenes (sin GUI para Railway)
opencv-python-headless==4.8.1.78
Pillow==10.1.0
numpy==1.24.4